requirements.txt
feedparser>=6.0.10
supabase>=2.0.0
PyYAML>=6.0
python-dateutil>=2.8.2
requests>=2.30.0
